Allow `topo::Builder::with_tips` and `with_ends` for non-default pred
We recently introduced the associated fns `with_tips` and `with_ends`
for adding tips and ends after initial creation of a `Builder`. We
erroneously added them to the one impl constraining the predicate to a
specific function pointer type for the default predicate. This meant
that tips and ends could only be added until the predicate is set to
a `FnMut` that can not be coerced to that same type.

This change lifts this limitation by moving `with_tips` and `with_ends`
to the more generic impl for `Builder`.