fix!: symlinks_to_directories_are_ignored_like_directories by value
The methods of `gix::dirwalk::Options` are paired, where for each
option `X` of `Options`, a method named like `X` takes and returns
`self` by value, and a method `set_X` takes and returns `self` by
mutable reference.

But in `symlinks_to_directories_are_ignored_like_directories`, both
took `self` by mutable reference. This fixes that. The effect of
this fix is to allow building `Options` with a call to that method
as the last factory method call (and using it where `Options` is
accepted but `&mut Options` is not).

Most code that consumes the crate should be unaffected, but:

- Code where calls were ordered unnaturally to avoid putting such
  a call last should be able to be improved.

- Code that used the method like its `set_*` countepart
  `set_symlinks_to_directories_are_ignored_like_directories` will
  be broken. That's what makes this fix a breaking change. Any such
  code can be fixed by modifying it to call the `set_*` version
  instead, which is probably what would have been intended anyway.