Remove the `test-32bit-cross` job
It was not able to test as much as was desired, and there is now a
job-level container based way that does a full 32-but ARMv7 test
(facilitated by the new ARM64 runners, which are also able to
execute ARMv7 binaries without emulation).

The new ARM jobs added in the last couple of commits, consisting
of a 64-bit job in `test-fast` and a 32-bit job in `test-32bit`,
increase the total time and resources required to run CI checks.
Removing `test-32bit-cross` only slightly counteracts that. But it
is also less needed now.