Use `ubuntu-22.04-arm`; clean up temporary changes
In the AArch64/ARM64 (64-bit, non-containerized) test-fast job,
this uses the `ubuntu-22.04-arm` runner instead of the
`ubuntu-24.04-arm` runner. This is to avoid the errors described
in #1790, i.e., to work around rust-lang/rust#135867.

Such problems have not been observed on the 22.04 runner, including
in tests intended to find them, and switching to it seems to be a
complete workaround for the problem. In contrast, continuing to use
the 24.04 runner, but attempting to work around the problem by
switching from the stable to the beta channel, looks like it would
greatly decrease the frequency of the errors but not eliminate
them. A problem with `actions/checkout` failing is likewise
observed on the 24.04 runner only, so using 22.04 avoids that too.

Because that seems like a complete workaround, this also reverts
50da7cb (#1792). That is to say that the ARM64 test-fast job is
again in the `test-fast` matrix. It is capable of cancelling or
being cancelled by the other `test-fast` checks. Code duplication
in the workflow is somewhat decreased. The job will again block PR
auto-merge.

Similar errors do not seem to have occurred in the `test-32bit`
job that runs an arm32v7 Docker image in `ubuntu-24.04-arm`, and it
is not clear that changing the runner image would help with #1780,
nor even if that issue is still happening. Therefore, it is not
changed there at this time.

This affects only ARM Linux runners. The x86-64 runners continue to
use `ubuntu-latest`, which is currently resolved to `ubuntu-24.04`,
and that does not need to be changed. Likewise, the `macos-latest`
runners use ARM processors (Apple Silicon) and they are fine.

Various experiments were done in a separate workflow. This commit
also removes that workflow, because it is not actively needed
anymore, and because, if kept, it would have to be modified to
avoid running hundreds of extra checks on each and every push.