doc: Fix `to_windows_separators` docstring, revise others
The `gix_path::convert::to_windows_separators` documentation
comment had inadvertently described the opposite of what this
function does, saying it changed backslashed to slashes and that
the effect was like a Unix path. This fixes the docstring so it
saye `to_windows_separators` changes slashes to backslashes, and
moves the text that likens it to paths on Unix to the
`to_unix_separators` docstring, where it applies.

This also makes a number of other more minor revisions for
clarity and stylistic consistency to documentation comments on
other functions in the `gix_path::convert` module.