Test other UNC-like prefixes besides `\\?\`
The test cases added in this commit are not very high value, since
any `\` in what is supposed to be a component should be detected as
invalid on Windows, which entails that these paths are detected as
such. Adding these is mainly to clarify that the `\\?\` is not
special here, comparedd to other kinds of `\` and `\\` paths.