Fix and enable disallow_shell test on Windows
The `disallow_shell` test case for `gix-command` has two parts. The
first part runs a command that would fail if not really run in a
shell, mostly to verify that the second part of the test is robust
enough. The second part actually tests the `Prepare::without_shell`
method.

While neither part should be made any less robust, only the first
part had failed on Windows, and the reason was (or had become) only
that it would not really run in a shell, due to being eligible for
manual argument splitting and `allow_manual_arg_splitting` being
`true` by default on Windows.

Therefore, this modifies the first part to use a method that sets
`allow_manual_arg_splitting` to `false`. This is sufficient to make
the test pass on Windows, so this also enables it on all platforms.