feat: Use `gix_path::env::shell()` in `gix-command` for sh
`gix_command::Prepare` previously used `sh` on Windows rather than
first checking for a usable `sh` implementation associated with a
Git for Windows installation. This changes it to use the `gix-path`
facility for finding what is likely the best 'sh' implementation
for POSIX shell scripts that will operate on Git repositories. This
increases consistency across how different crates find 'sh', and
brings the benefit of preferring the Git for Windows `sh.exe` on
Windows when it can be found reliably.