Fix Windows-related `-c` options in `match_baseline_files`
The `match_baseline_files.sh` fixture script adds entries to the
index using `git update-index --add --cacheinfo`. These entries
don't exist on disk, so it should be okay that some would cause
problems on Windows if an attempt were made to actually create
them. The tests that use the generated repository work on Windows.

It is intended that the fixture script also be able to run on
Windows, and Windows-related `-c` options are passed in commands
where they are expected to be needed: `core.ignoreCase=false` and
`core.protectNTFS=false`.

But the fixture script nonetheless was failing on Windows because
two commands that don't need `core.ignoreCase=false` but *do* need
`core.protectNTFS=false` had `core.ignorecase=false` instead. This
corrects that.

This fixes two of the long-standing `GIX_TEST_IGNORE_ARCHIVES=1`
Windows failures reported in #1358:

- gix-pathspec::pathspec search::files
- gix-pathspec::pathspec search::prefixes_are_always_case_sensitive